Move util function to utils.ts
Fix flaky tests by ensuring we only query by emojis in that test
