+++ src/typescript/sdk/tests/e2e/queries/sorted/utils.ts
@@ -1,26 +1,8 @@
-// Postgres doesn't deterministically sort when columns are equal. In order to compare the
-// results when values are equal, we must partition all results into groups based on their
-// values, and compare the differences in the sets of results, and then compare the earliest
-// "appearance" of any value in each `value: (...marketID)` set against the one in the
-// unsorted results set.
-//
-// That is, say we have { 1n: 100, 2n: 200, 3n: 200 } in the table, where `1n` is a marketID
-// and `100` is the query'es relevant value (the second field in the `mapFunction`).
-// In our first sorted query results (the one passed in), the resulting mapped array might be:
-// [ [marketID, value, index], ... ]
-// [ [1n, 100, 1], [2n, 200, 2], [3n, 300, 3] ]
-// OR
-// [ [1n, 100, 1], [3n, 200, 2], [2n, 200, 3] ]
-// Note that the market ID 3n appears before 2n, since they have the same result.
-//
-// Thus in order to compare the query results with subsets of unsorted market IDs (the ones
-// with equal `value` fields), we will bin each result according to its `value` field, then
-// get the lowest index in the bin. Then we compare that to the manually sorted result.
-// In the case above, this would be:
-// { 100: [ Set(1n), 1], 200: [ Set(2n, 3n), 2] }
-// Then when comparing each query's results in this structure, we can know what (if any)
-
-import { sortBigIntArrays } from "../../../../src";
+import { compareBigInt, sortBigIntArrays } from "../../../../src";
+import { postgrest, queryHelper } from "../../../../src/indexer-v2/queries";
+import { MarketStateModel, toMarketState } from "../../../../src/indexer-v2/types";
+import { TableName } from "../../../../src/indexer-v2/types/json-types";
+import { LIMIT } from "../../../../src/queries";
 
 // differences there are in the marketIDs, values and sort order.
 type SortableResults = {
@@ -46,13 +28,73 @@ export const sortWithUnsortedSubsets = (
   return valueMap;
 };
 
+// Market IDs are in ascending order of registration time, so to filter out markets registered
+// after a query is run, we can just get the latest market ID from that query's results.
+// Then we filter out markets with a market ID greater than that and sort the values in order to
+// compare the expected results with the actual query results.
+export const checkOrder = async (
+  queryResults: MarketStateModel[],
+  sort: (a: MarketStateModel, b: MarketStateModel) => 0 | -1 | 1,
+  mapFunction: (
+    value: MarketStateModel,
+    index: number
+  ) => { marketID: bigint; value: bigint; index: number }
+) => {
+  // If the query results are longer than `LIMIT`, these queries need to be paginated.
+  expect(queryResults.length).toBeLessThan(LIMIT);
+  const results = queryResults.map(mapFunction);
+
+  // Filter the unsorted query by the market IDs that are in the `queryResults`.
+  const marketIDsInQueryResult = queryResults.map(({ market }) => market.marketID);
+  // prettier-ignore
+  const unsortedQuery = () =>
+      postgrest
+        .from(TableName.MarketState)
+        .select("*")
+        .in("market_id", marketIDsInQueryResult);
+
+  // Manually sort the `unsortedQuery` results by the `sort` function passed in.
+  const expected = (await queryHelper(unsortedQuery, toMarketState)({}))
+    .toSorted(sort)
+    .map(mapFunction);
+
+  checkSubsetsEqual(results, expected);
+};
+
+// Postgres doesn't deterministically sort when columns are equal. In order to compare the
+// results when values are equal, we must partition all results into groups based on their
+// values, and compare the differences in the sets of results, and then compare the earliest
+// "appearance" of any value in each `value: (...marketID)` set against the one in the
+// unsorted results set.
+//
+// That is, say we have { 1n: 100, 2n: 200, 3n: 200 } in the table, where `1n` is a marketID
+// and `100` is the query'es relevant value (the second field in the `mapFunction`).
+// In our first sorted query results (the one passed in), the resulting mapped array might be:
+// [ [marketID, value, index], ... ]
+// [ [1n, 100, 1], [2n, 200, 2], [3n, 300, 3] ]
+// OR
+// [ [1n, 100, 1], [3n, 200, 2], [2n, 200, 3] ]
+// Note that the market ID 3n appears before 2n, since they have the same result.
+//
+// Thus in order to compare the query results with subsets of unsorted market IDs (the ones
+// with equal `value` fields), we will bin each result according to its `value` field, then
+// get the lowest index in the bin. Then we compare that to the manually sorted result.
+// In the case above, this would be:
+// { 100: [ Set(1n), 1], 200: [ Set(2n, 3n), 2] }
+// Then when comparing each query's results in this structure, we can know what (if any)
 export const checkSubsetsEqual = (results: SortableResults[], expected: SortableResults[]) => {
   expect(results.length).toEqual(expected.length);
   const binnedResults = sortWithUnsortedSubsets(results);
   const binnedExpected = sortWithUnsortedSubsets(expected);
+  const sortedResultKeys = Array.from(binnedResults.keys()).toSorted(compareBigInt);
+  const sortedExpectedKeys = Array.from(binnedExpected.keys()).toSorted(compareBigInt);
+  if (sortedResultKeys.join(",") !== sortedExpectedKeys.join(",")) {
+    console.log(binnedResults);
+    console.log(binnedExpected);
+    console.log(sortedResultKeys);
+    console.log(sortedExpectedKeys);
+  }
   expect(binnedResults.size).toEqual(binnedExpected.size);
-  const sortedResultKeys = Array.from(binnedResults.keys()).toSorted();
-  const sortedExpectedKeys = Array.from(binnedExpected.keys()).toSorted();
   expect(sortedResultKeys).toEqual(sortedExpectedKeys);
 
   for (const key of binnedResults.keys()) {
